+++ Tweed.Web/Views/Shared/_Tweed.cshtml.cs
@@ -9,6 +9,6 @@ public class TweedViewModel
     public string? Text { get; set; }
     public ZonedDateTime? CreatedAt { get; set; }
     public string? Id { get; set; }
-    public int? LikesCount { get; set; }
+    public long? LikesCount { get; set; }
     public bool LikedByCurrentUser { get; set; }
 }
